Move default values of plugin parameters to SettingsProjectState
